// Backup25+26-03-2021/src/app/model/course.ts
export class Course {
    courseId: Int32Array;
    courseName: string;
    courseDuration: string;    
    description: string;
    startDate: Date;
    lastUpdate: Date;
    hastag: string;
    level: string;
    thumbnailImage: string;
    rating: number;
    numberOfVoters: number;
    numberOfParticipants: number;
    price: number;
    lessonNumbers: number;
    isActive:boolean;
    accountId: Int32Array;
}